changed expires default
expires=0 will make sure a cookie is not stored. Not a nice default for a 'set'  command. If you don't specify expires in cfcookie it will stay valid for the current session.
That's what I added as a default now. It only add expires to cfcookie if it is defined in the arguments.